backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
Pillow>=10.2.0
pytesseract==0.3.10
requests==2.31.0
djangorestframework-simplejwt==5.3.0
